requirements.txt
streamlit~=1.41.1
jupyter~=1.1.1
pip~=24.3.1
attrs~=24.3.0
protobuf~=5.29.3
pillow~=11.1.0
tornado~=6.4.2
Jinja2~=3.1.5
streamlit~=1.41.1
pandas~=2.2.3
pydeck~=0.9.1
numpy~=2.2.2
requests~=2.32.3
python-dotenv~=1.0.1